pyproject.toml
[project]
name = "sam2-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sam-2 @ git+https://github.com/facebookresearch/sam2.git",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pillow>=9.0.0",
    "requests",
    "matplotlib",
    "opencv-python",
    "imageio",
    "imageio-ffmpeg",
    "ipykernel>=7.1.0",
    "tqdm",
    "sentencepiece",
    "protobuf",
    "scikit-learn>=1.0.0",
]
